enable NoTypeInfo test
