+++ src/test/java/com/fasterxml/jackson/databind/jsontype/NoTypeInfoTest.java
@@ -9,11 +9,37 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
 import com.fasterxml.jackson.databind.testutil.NoCheckSubTypeValidator;
 
+import java.util.Arrays;
+import java.util.List;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class NoTypeInfoTest extends DatabindTestUtil
 {
+    @JsonTypeInfo(use = JsonTypeInfo.Id.NAME)
+    static class Value1654 {
+        public int x;
+
+        protected Value1654() {
+        }
+
+        public Value1654(int x) {
+            this.x = x;
+        }
+    }
+
+    static class Value1654TypedContainer {
+        public List<Value1654> values;
+
+        protected Value1654TypedContainer() {
+        }
+
+        public Value1654TypedContainer(Value1654... v) {
+            values = Arrays.asList(v);
+        }
+    }
+
     @JsonTypeInfo(use=JsonTypeInfo.Id.NONE)
     @JsonDeserialize(as=NoType.class)
     static interface NoTypeInterface {
@@ -44,5 +70,20 @@ public class NoTypeInfoTest extends DatabindTestUtil
         assertNotNull(bean);
         NoType impl = (NoType) bean;
         assertEquals(6, impl.a);
+    }
+
+    // [databind#1654]
+    @Test
+    void noTypeElementOverride() throws Exception {
+        // regular typed case
+        final ObjectMapper mapper = newJsonMapper();
+        String json = mapper.writeValueAsString(new Value1654TypedContainer(
+            new Value1654(1),
+            new Value1654(2),
+            new Value1654(3)
+        ));
+        Value1654TypedContainer result = mapper.readValue(json, Value1654TypedContainer.class);
+        assertEquals(3, result.values.size());
+        assertEquals(2, result.values.get(1).x);
     }
 }
